Refactor: python, cts: Move SplitBrainTest into the _cts module.
No code changes.
